largest-contentful-paint/web-font-text-resize.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @font-face {
    font-family: 'TestFaceSmaller';
    src: url('/fonts/CSSTest/csstest-weights-15-w1.ttf?pipe=trickle(d0.5)');
    font-display: swap;
  }

  .test {
    font-family: 'TestFaceSmaller';
  }

</style>
<!--
  Web-font styled text that gets resized during swap period should not make a
  LCP emission if the new size is smaller than the current LCP element size.
-->
<div class="test">LCP: Web Font Styled Text Resize</div>
<script>
  function LCPEntryList(t) {
    return new Promise(resolve => {
      let = lcpEntries = [];
      new PerformanceObserver((entryList, observer) => {
        lcpEntries = lcpEntries.concat(entryList.getEntries());
        if (lcpEntries) {
          t.step_timeout(() => {
            resolve(lcpEntries);
            observer.disconnect();
          }, 200);
        }
      }).observe({ type: 'largest-contentful-paint', buffered: true });
    });
  }

  promise_test(async t => {
    await document.fonts.ready;

    // Verify web font is loaded.
    assert_true(!!window.PerformanceResourceTiming, "ResourceTiming not \
    supported");
    let url = '/fonts/CSSTest/csstest-weights-15-w1.ttf?pipe=trickle(d0.5)';
    var absoluteURL = new URL(url, location.href).href;
    assert_equals(performance.getEntriesByName(absoluteURL).length, 1, 'Web font\
    should be downloaded');

    // Verify web font has been applied.
    assert_equals(getComputedStyle(
      document.getElementsByClassName('test')[0]).fontFamily, 'TestFaceSmaller',
      'Font should be the web font added');

    // Verify there is only one LCP entry and it is still the one that
    // corresponds to the text of original font by checking its size.
    let entryList = await LCPEntryList(t);
    assert_equals(entryList.length, 1, 'Web font styled text resize that occurs\
    during swap period but is smaller should not make a new LCP emission.')

    assert_approx_equals(entryList[0].size, 4005, 300, 'The LCP entry should have size of \
    4218.');

  }, "LCP should be not updated if the web font styled text resizes to be \
  smaller during the swap period");
</script>
